saas_onboarding/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.22
sqlmodel>=0.0.8
sqlalchemy>=1.4
aiosqlite>=0.17
python-jose>=3.3.0
passlib[bcrypt]>=1.7
stripe>=5.0.0
jinja2>=3.1
python-multipart>=0.0.6
slowapi>=0.1.4
python-dotenv>=1.0.0
aiohttp>=3.9.5
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
